Update: to pull

+++ src/main/java/com/ddiring/backend_market/investment/dto/request/InvestmentRequest.java
@@ -13,6 +13,6 @@ public class InvestmentRequest {
 
     private String userSeq;
     private String projectId;
-    private Integer investedPrice;
+    private Long investedPrice;
     private Integer tokenQuantity;
 }
